split up functions and constants to own files
