+++ index.js
@@ -1,178 +1,18 @@
+// Import functions and const
+const { isAdmin, splitMessage, sendCmdResp, textToSpeech, formatDate, callOpenAIWithRetry, voiceEmbed } = require('./helpers');
+const { voiceDescriptions, voiceMapping, ne_voiceDescriptions, modelName, botCommand, DISABLED_MSG, COMMAND_PERM_MSG, ENABLE_MSG, DISABLE_MSG, RESET_MSG, DYNAMIC_RESET_MSG, RESET_ERROR_MSG, PERSONALITY_MSG, CASE_MODE, REPLY_MODE, BOT_REPLIES} = require('./constants');
 // Require the necessary discord.js classes
-const { Client, Discord, GatewayIntentBits, PermissionFlagsBits } = require('discord.js');
-const { EmbedBuilder } = require('discord.js');
+const { Client, EmbedBuilder, GatewayIntentBits } = require('discord.js');
+const fs = require('fs');
+const path = require('path');
 // Initialize .env config file
 require('dotenv').config();
-const botCommand = process.env.BOT_COMMAND;
 // Require openai and set API key and setup
 const { Configuration, OpenAIApi } = require("openai");
 const configuration = new Configuration({
 	apiKey: process.env.OPENAI_API_KEY,
 });
-const modelName = process.env.GPT_MODEL;
 const openai = new OpenAIApi(configuration);
-
-// TTS Related settings and functions
-// Require ffmpeg, fs, path, child_process for working with audio files
-const fs = require('fs');
-const path = require('path');
-const ffmpegPath = require('ffmpeg-static');
-const { exec } = require('child_process');
-// Require TikTok TTS package and store sessionID and URL
-const { config, createAudioFromText } = require('tiktok-tts');
-const sessionID = process.env.SESSION_ID;
-const BASE_URL = process.env.BASE_URL;
-config(sessionID, BASE_URL);
-// Voice mappings
-const voiceMapping = {
-	default: 'en_us_rocket',
-    au_female: 'en_au_001',
-    au_male: 'en_au_002',
-    uk_male1: 'en_uk_001',
-    uk_male2: 'en_uk_003',
-    us_female1: 'en_us_001',
-    us_female2: 'en_us_002',
-    us_male1: 'en_us_006',
-    us_male2: 'en_us_007',
-    us_male3: 'en_us_009',
-    us_male4: 'en_us_010',
-	alto: 'en_female_f08_salut_damour',
-    tenor: 'en_male_m03_lobby',
-    warmy_breeze: 'en_female_f08_warmy_breeze',
-    sunshine_soon: 'en_male_m03_sunshine_soon',
-    narrator: 'en_male_narration',
-    wacky: 'en_male_funny',
-    peaceful: 'en_female_emotional',
-    serious: 'en_male_cody',
-    pirate: 'en_male_pirate',
-    glorious: 'en_female_ht_f08_glorious',
-    funny_sing: 'en_male_sing_funny_it_goes_up',
-    chipmunk: 'en_male_m2_xhxs_m03_silly',
-    dramatic: 'en_female_ht_f08_wonderful_world',
-	ghostface: 'en_us_ghostface',
-    chewbacca: 'en_us_chewbacca',
-    c3po: 'en_us_c3po',
-    stitch: 'en_us_stitch',
-    stormtrooper: 'en_us_stormtrooper',
-    rocket: 'en_us_rocket',
-    fr_male1: 'fr_001',
-    fr_male2: 'fr_002',
-    de_female: 'de_001',
-    de_male: 'de_002',
-    es_male: 'es_002',
-    mx_male: 'es_mx_002',
-    br_female1: 'br_001',
-    br_female2: 'br_003',
-    br_female3: 'br_004',
-    br_male: 'br_005',
-    id_female: 'id_001',
-    jp_female1: 'jp_001',
-    jp_female2: 'jp_003',
-    jp_female3: 'jp_005',
-    jp_male: 'jp_006',
-    kr_male1: 'kr_002',
-    kr_female: 'kr_003',
-    kr_male2: 'kr_004'
-};
-
-// Voice descriptions
-const voiceDescriptions = {
-    au_female: 'English AU - Female',
-    au_male: 'English AU - Male',
-    uk_male1: 'English UK - Male 1',
-    uk_male2: 'English UK - Male 2',
-    us_female1: 'English US - Female 1',
-    us_female2: 'English US - Female 2',
-    us_male1: 'English US - Male 1',
-    us_male2: 'English US - Male 2',
-    us_male3: 'English US - Male 3',
-    us_male4: 'English US - Male 4',
-	alto: 'Alto',
-    tenor: 'Tenor',
-    warmy_breeze: 'Female',
-    sunshine_soon: 'Male',
-    narrator: 'Narrator',
-    wacky: 'Wacky',
-    peaceful: 'Female (UK)',
-    serious: 'Male',
-    pirate: 'Male',
-    glorious: 'Female Singing',
-    funny_sing: 'Rising Male Singing',
-    chipmunk: 'High Pitched Singing',
-    dramatic: 'Female Singing',
-	ghostface: 'Ghost Face',
-    chewbacca: 'Chewbacca',
-    c3po: 'C3PO',
-    stitch: 'Stitch',
-    stormtrooper: 'Stormtrooper',
-    rocket: 'Rocket (Guardians)'
-};
-// Voice descriptions
-const ne_voiceDescriptions = {
-    fr_male1: 'French - Male 1',
-    fr_male2: 'French - Male 2',
-    de_female: 'German - Female',
-    de_male: 'German - Male',
-    es_male: 'Spanish - Male',
-    mx_male: 'Spanish MX - Male',
-    br_female1: 'Portuguese BR - Female 1',
-    br_female2: 'Portuguese BR - Female 2',
-    br_female3: 'Portuguese BR - Female 3',
-    br_male: 'Portuguese BR - Male',
-    id_female: 'Indonesian - Female',
-    jp_female1: 'Japanese - Female 1',
-    jp_female2: 'Japanese - Female 2',
-    jp_female3: 'Japanese - Female 3',
-    jp_male: 'Japanese - Male',
-    kr_male1: 'Korean - Male 1',
-    kr_female: 'Korean - Female',
-    kr_male2: 'Korean - Male 2'
-};
-
-async function mergeAudioFiles(files, outputFile) {
-    const currentDirectory = __dirname;
-    const filterString = files.map((_, index) => `[${index}:0]`).join('');
-    const inputFiles = files.map(file => `-i "${path.join(currentDirectory, file)}"`).join(' ');
-
-    const outputFilePath = path.join(currentDirectory, outputFile);
-    
-    return new Promise((resolve, reject) => {
-        exec(`${ffmpegPath} ${inputFiles} -filter_complex "${filterString}concat=n=${files.length}:v=0:a=1[out]" -map "[out]" "${outputFilePath}"`, (error) => {
-            if (error) {
-                reject(error);
-            } else {
-                resolve();
-            }
-        });
-    });
-}
-
-async function textToSpeech(speaker, text) {
-    const parts = splitMessage(text, 250);
-    const audioFiles = await Promise.all(parts.map(async part => {
-        const fileName = `${Date.now()}_${speaker}_tts`;
-        await createAudioFromText(part, fileName, speaker);
-        return fileName.replace('_tts', '_tts.mp3');
-    }));
-
-    if (audioFiles.length === 1) {
-        return audioFiles[0];
-    } else {
-        const mergedFilename = `${Date.now()}_${speaker}_tts_merged.mp3`;
-        await mergeAudioFiles(audioFiles, mergedFilename);
-
-        // Clean up temporary audio files
-        audioFiles.forEach(file => {
-            fs.unlink(file, (err) => {
-                if (err) {
-                    console.error(err);
-                }
-            });
-        });
-
-        return mergedFilename;
-    }
-}
 
 // Create a new Discord client instance
 const client = new Client({intents: [GatewayIntentBits.Guilds,GatewayIntentBits.GuildMessages,GatewayIntentBits.MessageContent,] });
@@ -208,6 +48,7 @@ function initPersonalities() {
 // Run function
 initPersonalities();
 
+// Get current personality from message
 function getPersonality(message) {
     let personality = null;
     let earliestIndex = Infinity;
@@ -227,143 +68,29 @@ function getPersonality(message) {
     // Return the personality of the message
     return personality;
 }
-
-// Split message function
-function splitMessage(resp, charLim) {
-	const responseNum = Math.ceil(resp.length / charLim);
-	const responses = new Array(responseNum);
-	// For the number of split responses, if its the last response, make the size the character limit, else make the size the last index of a space that is under 2000 characters
-	for (let i = 0, c = 0, chunkSize = null; i < responseNum; i++, c+=chunkSize) {
-		if (i + 1 >= responseNum) {
-			chunkSize = charLim;
-		} else {
-					chunkSize = resp.substr(c, charLim).lastIndexOf(" ");
-		}
-		responses[i] = resp.substr(c, chunkSize);
-	}
-	return responses;
-}
-
-// Send command responses function
-function sendCmdResp(msg, cmdResp) {
-    if (process.env.REPLY_MODE === 'true') {
-        msg.reply(cmdResp);
-    } else {
-        msg.channel.send(cmdResp);
-    }
-}
-
 // Create Pause var
 client.isPaused = false;
 
-// Set admin user IDs
-adminId = process.env.ADMIN_ID.split(',');
-
-// Check message author id function
-function isAdmin(msg) {
-	if (msg.member.permissions.has(PermissionFlagsBits.Administrator) || msg.member.permissions.has(PermissionFlagsBits.ManageMessages)) {
-		return true;
-	} else {
-		return adminId.includes(msg.author.id);
-	}
-}
-function formatDate(date) {
-	const year = date.getFullYear();
-	const month = String(date.getMonth() + 1).padStart(2, '0');
-	const day = String(date.getDate()).padStart(2, '0');
-	const hours = String(date.getHours()).padStart(2, '0');
-	const minutes = String(date.getMinutes()).padStart(2, '0');
-	const seconds = String(date.getSeconds()).padStart(2, '0');
-	return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-}
-async function callOpenAIWithRetry(apiCall, retries, delay) {
-	let attempts = 0;
-  
-	while (attempts < retries) {
-	  try {
-		const response = await apiCall();
-		return response;
-	  } catch (error) {
-		if (error.statusCode === 502 && attempts < retries - 1) {
-		  console.log('Encountered a 502 error. Retrying...');
-		  attempts++;
-		  await new Promise((resolve) => setTimeout(resolve, delay));
-		} else {
-		  throw error;
-		}
-	  }
-	}
-}
-function groupVoiceDescriptions(voiceDescriptions) {
-    const grouped = {};
-    const regex = /^(.*\D)(\d+)$/;
-
-    for (const key in voiceDescriptions) {
-        const match = key.match(regex);
-        if (match) {
-            const prefix = match[1];
-            const number = match[2];
-
-            if (grouped[prefix]) {
-                grouped[prefix].push(number);
-            } else {
-                grouped[prefix] = [number];
-            }
-        } else {
-            grouped[key] = voiceDescriptions[key];
-        }
-    }
-
-    const newVoiceDescriptions = {};
-    for (const key in grouped) {
-        if (Array.isArray(grouped[key])) {
-            const newKey = `${key}${grouped[key].join('_')}`;
-            newVoiceDescriptions[newKey] = voiceDescriptions[`${key}${grouped[key][0]}`];
-        } else {
-            newVoiceDescriptions[key] = grouped[key];
-        }
-    }
-
-    return newVoiceDescriptions;
-}
-
-function voiceEmbed(voiceDescriptions, title) {
-	const formattedDescriptions = groupVoiceDescriptions(voiceDescriptions);
-	// create an embed object
-	let voiceEmbed = new EmbedBuilder()
-		.setColor(0x0099FF) // set the color of the embed
-		.setTitle(title) // set the title of the embed
-		.setDescription('Here are the '+ title +' you can use'); // set the description of the embed
-
-	// loop through your voiceDescriptions object and add fields to the embed
-	for (let key in formattedDescriptions) {
-		const formattedKey = key.replace(/(\d)_/g, '$1,');
-		voiceEmbed.addFields({ name: formattedKey, value: formattedDescriptions[key], inline: true });
-	}
-
-	return voiceEmbed;
-}
-
 client.on('messageCreate', async msg => {
 	// Don't do anything when message is from the bot itself or from other bots, based on the BOT_REPLIES environment variable
-	if (msg.author.id === client.user.id || (msg.author.bot && process.env.BOT_REPLIES !== "true")) return;
+	if (msg.author.id === client.user.id || (msg.author.bot && BOT_REPLIES !== "true")) return;
 
 	// Enable/Disable bot commands
 	if (msg.content === botCommand + 'disable') {
 		if (isAdmin(msg)) {
 			client.isPaused = true;
-			sendCmdResp(msg, process.env.DISABLE_MSG);
+			sendCmdResp(msg, DISABLE_MSG);
 		} else {
-			sendCmdResp(msg, process.env.COMMAND_PERM_MSG);
+			sendCmdResp(msg, COMMAND_PERM_MSG);
 			return;
 		}
 	}
 	if (msg.content === botCommand + 'enable') {
 		if (isAdmin(msg)) {
 			client.isPaused = false;
-			sendCmdResp(msg, process.env.ENABLE_MSG);
+			sendCmdResp(msg, ENABLE_MSG);
 		} else {
-			sendCmdResp(msg, process.env.COMMAND_PERM_MSG);
+			sendCmdResp(msg, COMMAND_PERM_MSG);
 			return;
 		}
 	}
@@ -372,14 +99,14 @@ client.on('messageCreate', async msg => {
 	if (msg.content.startsWith(botCommand + 'reset')) {
 		// Check disabled status
 		if (client.isPaused === true && !isAdmin(msg)) {
-			sendCmdResp(msg, process.env.DISABLED_MSG);
+			sendCmdResp(msg, DISABLED_MSG);
 			return;
 		}
 		let cutMsg = msg.content.slice(7);
 		// Delete all memories if message is "!reset all"
 		if (cutMsg === 'all') {
 			initPersonalities();
-			sendCmdResp(msg, process.env.RESET_MSG);
+			sendCmdResp(msg, RESET_MSG);
 			return;
 		} else {
 			// Check what personality's memory to delete
@@ -388,12 +115,12 @@ client.on('messageCreate', async msg => {
 				if (cutMsg.toUpperCase().startsWith(thisPersonality.name.toUpperCase())) {
 					let originalSystemMessage = thisPersonality.request.find(msg => msg.role === 'system');
 					personalities[i] = { "name": thisPersonality.name, "request" : [originalSystemMessage]};
-					sendCmdResp(msg, process.env.DYNAMIC_RESET_MSG.replace('<p>', thisPersonality.name));
+					sendCmdResp(msg, DYNAMIC_RESET_MSG.replace('<p>', thisPersonality.name));
 					return;
 				}
 			}
 			// Return error if reset message does not match anything
-			sendCmdResp(msg, process.env.RESET_ERROR_MSG);
+			sendCmdResp(msg, RESET_ERROR_MSG);
 			return;
 		}
 	}
@@ -402,13 +129,13 @@ client.on('messageCreate', async msg => {
 	if (msg.content === botCommand + 'personalities') {
 		// Check disabled status
 		if (client.isPaused === true && !isAdmin(msg)) {
-		  sendCmdResp(msg, process.env.DISABLED_MSG);
+		  sendCmdResp(msg, DISABLED_MSG);
 		  return;
 		}
 		// Create an embed object
 		let persEmbed = new EmbedBuilder()
 		  .setColor(0x0099FF) // set the color of the embed
-		  .setTitle(process.env.PERSONALITY_MSG) // set the title of the embed
+		  .setTitle(PERSONALITY_MSG) // set the title of the embed
 		  .setDescription('Here are some personalities and their prompts'); // set the description of the embed
 	  
 		// Add personality names and prompts to fields
@@ -480,11 +207,11 @@ client.on('messageCreate', async msg => {
 	if (msg.content.startsWith(botCommand + 'tts')) {
         // Check disabled status
         if (client.isPaused === true && !isAdmin(msg)) {
-            sendCmdResp(msg, process.env.DISABLED_MSG);
+            sendCmdResp(msg, DISABLED_MSG);
             return;
         }
 
-        const regex = /^!tts\s+(\S+)\s+(.+)/;
+        const regex = /^[^a-zA-Z]+tts\s+(\S+)\s+(.+)/;
         const matches = msg.content.match(regex);
 
         // Check for correct command format
@@ -613,7 +340,7 @@ client.on('messageCreate', async msg => {
 
 	// Check if bot disabled/enabled
 	if (client.isPaused === true && !isAdmin(msg)) {
-		sendCmdResp(msg, process.env.DISABLED_MSG);
+		sendCmdResp(msg, DISABLED_MSG);
 		return;
 	}
 
@@ -634,7 +361,7 @@ client.on('messageCreate', async msg => {
 		const responseChunks = splitMessage(response, 2000)
 		// Send the split API response
 		for (let i = 0; i < responseChunks.length; i++) {
-			if (process.env.REPLY_MODE === 'true' && i === 0) {
+			if (REPLY_MODE === 'true' && i === 0) {
 				msg.reply(responseChunks[i]);
 			} else {
 				msg.channel.send(responseChunks[i]);
@@ -661,7 +388,7 @@ async function chat(requestX){
 		let responseContent;
 
 		// Check capitlization mode
-		switch (process.env.CASE_MODE) {
+		switch (CASE_MODE) {
 			case "":
 				responseContent = completion.data.choices[0].message.content;
 				break;
